Revert "full double"
This reverts commit f383bd72ecf951256e432b26c6b8d16812c25814.

+++ src/solver/variable/include/antares/solver/variable/storage/stdDeviation.h
@@ -241,7 +241,7 @@ private:
         break;
         case Category::annual:
         {
-            const double d = (double)*array - (double)results.avgdata.year * (double)results.avgdata.year;
+            const double d = *array - (double)results.avgdata.year * (double)results.avgdata.year;
             *target = squareRootChecked(d);
         }
         break;
